pyproject.toml
[tool.poetry]
name = "trustwork-backend"
version = "0.1.0"
description = ""
authors = ["Swapnil Chopra <swapnil.chopra@webskitters.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.1.1"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.8"
geopy = "^2.4.1"
requests = "^2.32.3"
firebase-admin = "^6.6.0"
channels = "^4.2.0"
channels-redis = "^4.2.1"
daphne = "^4.1.2"
uvicorn = "^0.34.0"
websockets = "^14.1"
basicauth = "^1.0.0"
django-rest-swagger = "^2.2.0"
dj-stripe = "^2.9.0"
stripe = "^11.4.1"
pycountry = "^24.6.1"
twilio = "^9.4.5"
pyclean = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


